backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# CORS
python-dotenv==1.0.0

# WebSocket
websockets==12.0
python-socketio==5.10.0

# MongoDB
pymongo==4.6.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Date handling
python-dateutil==2.8.2

# Async
aiofiles==23.2.1

# Redis (optional for caching)
redis==5.0.1

# Encryption
cryptography==41.0.7